backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
celery>=5.3.6
redis>=5.0.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.0.0
channels>=4.0.0
asgiref>=3.7.0 